requirements_ubuntu.txt
# Eden Trading Bot - Ubuntu Server Requirements
# Simplified for Wine MT5 environment

numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0
requests>=2.25.0
python-dotenv>=0.19.0
scikit-learn>=1.0.0
xgboost>=1.5.0
ta>=0.10.0
schedule>=1.1.0
colorama>=0.4.4
tabulate>=0.8.9
